Renaming and adding instructions for loading and evaluation.
