+++ pfns/scripts/tabpfn_interface.py
@@ -6,7 +6,6 @@ from torch.utils.checkpoint import checkpoint
 
 from pfns.utils import normalize_data, to_ranking_low_mem, remove_outliers
 from pfns.utils import NOP, normalize_by_used_features_f
-# from tabpfn.utils import NOP, normalize_by_used_features_f
 
 from sklearn.preprocessing import PowerTransformer, QuantileTransformer, RobustScaler
 
@@ -18,7 +17,7 @@ from sklearn.utils.multiclass import check_classification_targets
 from sklearn.utils import column_or_1d
 from sklearn.preprocessing import LabelEncoder
 from pathlib import Path
-from pfns.scripts.model_builder import load_model_only_inference
+from pfns.scripts.tabpfn_model_builder import load_model_only_inference
 import os
 import pickle
 import io
